access-to-mysql/hello_world/requirements.txt
mysql-connector-python==8.0.18
aws-xray-sdk==2.4.2